+++ src/main/java/levi9/NotificationService/api/NotificationServiceApi.java
@@ -5,5 +5,5 @@ import org.springframework.cloud.openfeign.FeignClient;
 
 @FeignClient(name="notification-service")
 public interface NotificationServiceApi {
-    
+
 }
